moved several tests together
