+++ tests/test_z_integration.py
@@ -4,10 +4,11 @@ import os
 import pytest
 from schematic_db.rdb import MySQLDatabase
 from schematic_db.rdb.postgres import PostgresDatabase
+from schematic_db.rdb.synapse_database import SynapseDatabase
 from schematic_db.rdb_builder import RDBBuilder
 from schematic_db.schema import Schema
 from schematic_db.rdb_updater import RDBUpdater
-from schematic_db.manifest_store import ManifestStore, ManifestStoreConfig
+from schematic_db.manifest_store import ManifestStore
 from schematic_db.query_store import QueryStore
 from schematic_db.rdb_queryer import RDBQueryer
 
@@ -32,22 +33,14 @@ def fixture_rdb_builder_postgres(
     obj.rdb.drop_all_tables()
 
 
-@pytest.fixture(scope="module", name="manifest_store")
-def fixture_manifest_store(
-    test_synapse_project_id: str,
-    test_synapse_asset_view_id: str,
-    secrets_dict: dict,
-    test_schema_json_url: str,
+@pytest.fixture(scope="module", name="rdb_builder_synapse")
+def fixture_rdb_builder_synapse(
+    synapse_database: SynapseDatabase, test_schema2: Schema
 ) -> Generator:
-    """Yields a ManifestStore object"""
-    yield ManifestStore(
-        ManifestStoreConfig(
-            test_schema_json_url,
-            test_synapse_project_id,
-            test_synapse_asset_view_id,
-            secrets_dict["synapse"]["auth_token"],
-        )
-    )
+    """Yields a RDBBuilder with a synapse database and test schema"""
+    obj = RDBBuilder(rdb=synapse_database, schema=test_schema2)
+    yield obj
+    synapse_database.delete_all_tables()
 
 
 @pytest.fixture(scope="module", name="rdb_updater_mysql")
@@ -68,6 +61,16 @@ def fixture_rdb_updater_postgres(
     obj = RDBUpdater(rdb=postgres_database, manifest_store=manifest_store)
     yield obj
     obj.rdb.drop_all_tables()
+
+
+@pytest.fixture(scope="module", name="rdb_updater_synapse")
+def fixture_rdb_updater_synapse(
+    synapse_database: SynapseDatabase, manifest_store: ManifestStore
+) -> Generator:
+    """Yields a RDBUpdater with a synapse database and test schema"""
+    obj = RDBUpdater(rdb=synapse_database, manifest_store=manifest_store)
+    yield obj
+    synapse_database.delete_all_tables()
 
 
 @pytest.fixture(scope="function", name="query_store")
@@ -108,10 +111,10 @@ def fixture_rdb_queryer_postgres(
 
 
 @pytest.mark.schematic
-class TestRDBQueryer:
-    """Testing for RDBBuilder and test schema"""
+class TestIntegration:
+    """Integration tests"""
 
-    def test_mysql( #pylint: disable=too-many-arguments
+    def test_mysql(  # pylint: disable=too-many-arguments
         self,
         rdb_builder_mysql: RDBBuilder,
         rdb_updater_mysql: RDBUpdater,
@@ -122,6 +125,8 @@ class TestRDBQueryer:
         """Creates the test database in MySQL"""
         rdb_builder = rdb_builder_mysql
         assert rdb_builder.rdb.get_table_names() == []
+        rdb_builder.build_database()
+        assert rdb_builder.rdb.get_table_names() == test_schema_table_names
         rdb_builder.build_database()
         assert rdb_builder.rdb.get_table_names() == test_schema_table_names
 
@@ -137,7 +142,7 @@ class TestRDBQueryer:
         rdb_queryer.store_query_results(path)
         assert rdb_queryer.query_store.get_table_names() == test_schema_table_names
 
-    def test_postgres( #pylint: disable=too-many-arguments
+    def test_postgres(  # pylint: disable=too-many-arguments
         self,
         rdb_builder_postgres: RDBBuilder,
         rdb_updater_postgres: RDBUpdater,
@@ -148,6 +153,8 @@ class TestRDBQueryer:
         """Creates the test database in Postgres"""
         rdb_builder = rdb_builder_postgres
         assert rdb_builder.rdb.get_table_names() == []
+        rdb_builder.build_database()
+        assert rdb_builder.rdb.get_table_names() == test_schema_table_names
         rdb_builder.build_database()
         assert rdb_builder.rdb.get_table_names() == test_schema_table_names
 
@@ -162,3 +169,23 @@ class TestRDBQueryer:
         path = os.path.join(data_directory, "test_queries_postgres.csv")
         rdb_queryer.store_query_results(path)
         assert rdb_queryer.query_store.get_table_names() == test_schema_table_names
+
+    def test_synapse_update_all_database_tables(
+        self,
+        rdb_updater_synapse: RDBUpdater,
+        rdb_builder_synapse: RDBBuilder,
+        test_schema_table_names: list[str],
+    ) -> None:
+        """Creates the test database in Synapse"""
+        rdb_builder = rdb_builder_synapse
+        assert rdb_builder.rdb.get_table_names() == []
+        rdb_builder.build_database()
+        assert rdb_builder.rdb.get_table_names() == test_schema_table_names
+        rdb_builder.build_database()
+        assert rdb_builder.rdb.get_table_names() == test_schema_table_names
+
+        rdb_updater = rdb_updater_synapse
+        rdb_updater.update_database()
+        for name in test_schema_table_names:
+            table = rdb_updater.rdb.query_table(name)
+            assert len(table.index) > 0
